requirements.txt
# Production dependencies for the kitchen application
# Install with: pip install -r requirements.txt

# Web framework and server
fastapi>=0.114.0        # FastAPI framework; 0.114+ uses explicit ASGITransport (fixes httpx deprecation)
uvicorn==0.27.0         # ASGI server for FastAPI

# Authentication and security
PyJWT>=2.12.0           # JSON Web Tokens for authentication (CVE-2026-32597 fixed in 2.12.0)
passlib[bcrypt]==1.7.4  # Password hashing
bcrypt>=3.2.0,<4.1.0    # Passlib reads __about__; bcrypt 4.1+ removed it (AttributeError)

# Database
psycopg2-binary>=2.9.11  # PostgreSQL adapter; 2.9.11+ for Python 3.14

# Data validation and serialization
pydantic[email]>=1.10   # Data validation with email support
pydantic-settings>=2.0  # BaseSettings for env config (moved from pydantic in v2)
python-multipart>=0.0.9 # Form data parsing (fixes import deprecation)

# Rate limiting
slowapi>=0.1.9          # Rate limiting for unauthenticated endpoints

# Configuration and utilities
python-dotenv==1.0.1    # Environment variable loader
pyyaml                  # YAML configuration parsing
pytz==2024.1           # Timezone handling
pycountry==24.6.1      # ISO 3166 country code to name (alpha-2; offline, no API key)
phonenumbers>=8.13.0   # E.164 validation and normalization (offline, no API calls)

# HTTP client (for external API calls)
httpx>=0.27.0           # Async HTTP client (ASGI, WebSockets)
rapidfuzz>=3.0.0        # Fuzzy string matching (ingredient enrichment cron)
requests>=2.33.0        # HTTP requests library (CVE-2024-35195, CVE-2024-47081, CVE-2026-25645)
urllib3>=2.6.3          # HTTP client (CVE-2025-50181, CVE-2025-66418, CVE-2025-66471, CVE-2026-21441)

# QR Code generation
qrcode[pil]==7.4.2     # QR code generation with PIL support
Pillow>=12.0.0         # Python Imaging Library; 12.0+ for Python 3.14 compatibility

# GCS (Google Cloud Storage)
google-cloud-storage>=2.14.0
google-auth>=2.28.0     # Explicit: GCS signed URL impersonation (google.auth.impersonated_credentials)

# Stripe (subscription payments, test mode: sk_test_/pk_test_)
stripe>=8.0.0,<9.0.0

# SendGrid (production email; only imported when EMAIL_PROVIDER=sendgrid)
sendgrid>=6.11.0

# Jinja2 (email template rendering)
Jinja2>=3.1.0

# Ads platform: async job queue + Redis
arq>=0.26                   # Async Redis Queue (ARQ) for deferred conversion uploads
redis[hiredis]>=5.0         # Redis client with C parser (ARQ backend)

# Ads platform: GCP Secret Manager (runtime credential fetch)
google-cloud-secret-manager>=2.18.0

# Ads platform: Google Ads API (conversion uploads, campaign management)
google-ads>=24.0.0

# Ads platform: Meta Marketing API + Conversions API (CAPI)
facebook-business>=25.0.0
